requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
autopep8==1.4.3
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
conda==4.6.14
cryptography==2.4.2
cycler==0.10.0
Flask==1.0.2
future==0.17.1
geographiclib==1.49
geopy==1.19.0
googlemaps==3.0.2
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
menuinst==1.4.14
numpy==1.16.1
pep8==1.7.1
Pillow==6.0.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pygame==1.9.4
pylint==2.2.2
pyOpenSSL==18.0.0
pyparsing==2.3.1
PySocks==1.6.8
python-dateutil==2.8.0
python-telegram-bot==11.1.0
pywin32==223
requests==2.21.0
rope==0.12.0
ruamel-yaml==0.15.46
six==1.12.0
urllib3==1.24.1
Werkzeug==0.15.1
win-inet-pton==1.0.1
wincertstore==0.2
wrapt==1.11.1
